src/hyperliquid_python/requirements.txt
hyperliquid-python-sdk
pydantic>=2.10
requests
